RSA: Split PKCS#1 into its own submodule.
